Develop -> Main

+++ src/main/java/com/soma/backend/domain/report/entity/AmountRange.java
@@ -3,7 +3,7 @@ package com.soma.backend.domain.report.entity;
 /**
  * 청구·제시 금액 범위 값 객체. 확정 금액을 직접 노출하지 않고 범위·headroom만 계산한다(금소법 §17).
  */
-public record AmountRange(Long claimedMinAmount, Long claimedMaxAmount, Long offeredAmount) {
+public record AmountRange(Integer claimedMinAmount, Integer claimedMaxAmount, Integer offeredAmount) {
 
   /** claimed_max - offered. 음수(초과 제시) 가능. null 안전. */
   public long offerHeadroom() {
